requirements.txt
Django>=3.2.0,<4.0.0
djangorestframework>=3.12.0,<4.0.0
django-cors-headers>=3.7.0,<4.0.0
pandas>=1.3.0,<2.0.0
spacy>=3.0.0,<4.0.0
scikit-learn>=1.0.0,<2.0.0
pdfplumber>=0.7.0,<1.0.0
requests>=2.25.0,<3.0.0
beautifulsoup4>=4.9.0,<5.0.0 